feat(guard): restructure Spend + retire /theguard Insights view
/theguard (Guard Overview page):
- Retired the Overview / Insights view toggle. The "Insights" label was
  vague — same team-wide metrics live better on the new /theguard/spend/
  glance surface where they're grouped with a currency picker and month
  picker.
- Replaced the toggle with Overview (this page) + a Link to
  /theguard/spend/glance labelled "Spend at a glance."
- Deleted the entire {view === "insights" && <>} block, the insightsStats
  memo, the view + insightsPeriod state, and the loadEvents-on-view
  effect. Net -270 lines from the file.
- Backend app/routers/insights.py untouched — Lens tool definitions can
  still call it.

/theguard/spend (Spend section):
- SPEND_TABS = [Configure, Optimization]. The old "Overview" tab mixed
  budget controls with metric dashboards; splitting makes each tab do
  one job.
- /theguard/spend now shows ONLY the Spend controls card (team monthly
  budget, per-developer limit, alert threshold, hard cap). ~940-line
  file collapsed to 82 lines.
- /theguard/spend/glance (new URL) shows everything that used to live
  below Spend controls: currency + month picker, 4 top stat cards, RTK
  + Booster savings breakdown, By-developer table, By-AI-tool table.
- No SPEND_TABS on /theguard/spend/glance — it's a peer of Overview,
  not a Spend sub-tab. Guard rail still highlights Spend section while
  viewing it (activePrefixes match).

Shared plumbing extracted to apps/web/src/features/guard/spend/:
- shared.ts — Currency/Rate/Symbol constants, fromUsd/toUsd/
  formatTokens/formatMonthLabel helpers, TOOL_LABELS map, response
  shapes (SpendData / DeveloperSpend / AiToolBreakdown /
  TeamBudgetSettings).
- useSpendState.ts — single hook that owns team + role + workspace +
  currency + month + spend data + budgets + team settings + savings +
  saveTeamSettings + saveBudget + load. Both pages consume it. No
  duplicated fetch logic across Configure and Glance.
- components.tsx — SpendControlsPanel + BudgetBar + BudgetInput +
  MonthPicker, moved out of the old monolith so both pages import from
  one place.

Net stats: +3 new files (shared + hook + components), 2 page files
(Configure at 82 lines / Glance at 298 lines), old spend/page.tsx down
from 966 to 82 lines, theguard/page.tsx down 270 lines. Total diff
across changed files: -1152 / +820. No behavior change beyond the
tab / URL restructure.
